Fix the background scheduler to reschedule itself
The background thread would only reschedule itself with a potentially
updated configuration if it ever received a timeout. This would never
happen as the initial timeout was Duration::MAX, and on a configuration
update, we wouldn't recompute the next timeout. With this change, we
always compute the next wake up, and reset any pending wake up when
invalidating the configuration.

This fixes an issue where the background worker never performs any work.

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>